// types/enums.ts
export enum StockStatus {
  InStock = 0,
  OutOfStock = 1,
}

export enum Condition {
  New = 0,
  Used = 1,
  LikeNew = 2,
}

export enum FacetTypeEnum {
  CheckboxList = 0,
  RadioButtonList = 1,
  RangeSlider = 2,
  NumericInput = 3,
  BooleanSwitch = 4,
  SearchBox = 5,
  DateRangePicker = 6,
}